backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pyodbc==5.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1